+++ libs/ymir-core/include/ymir/db/game_db.hpp
@@ -47,6 +47,7 @@ struct GameInfo {
         FastBusTimings = 1ull << 4ull,        ///< Fast bus timings required to fix stability issues
         FastMC68EC000 = 1ull << 5ull,         ///< Overclocked MC68EC000 required to fix stability issues
         StallVDP1OnVRAMWrites = 1ull << 6ull, ///< Stall/slow down VDP1 drawing on VDP1 VRAM writes
+        SlowVDP1 = 1ull << 7ull,              ///< Slow down VDP1 processing overall
     };
 
     Flags flags = Flags::None;        ///< Game compatibility flags